Extract ClickStack API models and finalize library split
